Add files via upload
Test Computers Online is the new script that I have been working on for the last two days. This script shows you which computers are online and offline. For online computers, it tells you who a login user is.
More updates in the coming days as I am still thinking on how to make the script more efficient and useful.